Import of ASNMTAP/ASNMTAP-3.000015 from CPAN.
gitpan-cpan-distribution: ASNMTAP
gitpan-cpan-version:      3.000015
gitpan-cpan-path:         ASNMTAP/ASNMTAP-3.000015.tar.gz
gitpan-cpan-author:       ASNMTAP
gitpan-cpan-maturity:     released

+++ applications/sbin/perfparse_crontab.sh
@@ -2,7 +2,7 @@
 # ----------------------------------------------------------------------------------------------------------
 # © Copyright 2003-2007 Alex Peeters [alex.peeters@citap.be]
 # ----------------------------------------------------------------------------------------------------------
-# 2007/06/10, v3.000.014, perfparse_crontab.sh
+# 2007/10/21, v3.000.015, perfparse_crontab.sh
 # ----------------------------------------------------------------------------------------------------------
 
 if [ -f ~/.profile ]; then
@@ -13,7 +13,13 @@ if [ -f ~/.bash_profile ]; then
   source ~/.bash_profile
 fi
 
-logFilename="/opt/asnmtap/log/perfdata-asnmtap.log"
+AMPATH=/opt/asnmtap-3.000.xxx
+
+if [ "$ASNMTAP_PATH" ]; then
+  AMPATH=$ASNMTAP_PATH
+fi
+
+logFilename="$AMPATH/log/perfdata-asnmtap.log"
 
 if [ -e "$logFilename" ]; then
   if [ -e "/usr/local/bin/date" ]; then
@@ -27,10 +33,18 @@ if [ -e "$logFilename" ]; then
   fi
 
   epochFilename="$logFilename-$epoch"
+  epochFilenameFailed="$epochFilename-failed"
 
-  echo "Filenames: '$logFilename', '$epochFilename'";
+  echo "Filenames: '$logFilename', '$epochFilename', '$epochFilenameFailed'"
   mv $logFilename $epochFilename
-  cat $epochFilename | /opt/asnmtap/perfparse/bin/perfparse-log2mysql
-  rm $epochFilename
+  cat $epochFilename | $AMPATH/perfparse/bin/perfparse-log2mysql
+  rv="$?"
+
+  if [ "$rv" = "0" ]; then
+    rm $epochFilename
+  else
+    mv $epochFilename $epochFilenameFailed
+  fi
 fi
 
+exit $rv
